third_party/blink/web_tests/fast/css/fontfaceset-check-platform-fonts.html

<!DOCTYPE html>
<script src='../../resources/testharness.js'></script>
<script src='../../resources/testharnessreport.js'></script>
<script>

test(() => {
  assert_true(document.fonts.check('10px Arial'), 'Arial');
  assert_true(document.fonts.check('10px Nonexistent'), 'Nonexistent');
  assert_true(document.fonts.check('10px sans-serif'), 'sans-serif');
  assert_true(document.fonts.check('10px Nonexistent, monospace'), 'Nonexistent, monospace');
  assert_true(document.fonts.check('10px Nonexistent1, Nonexistent2'), 'Nonexistent1, Nonexistent2');
}, 'Tests FontFaceSet#check() returns true for platform fonts');

</script>
